requirements.txt
fastapi
uvicorn
sqlalchemy
pyTigerGraph
python-dotenv
